Update for mem-telstate branch of katsdptelstate
Needed once ska-sa/katsdptelstate#71 lands.

It was direct accessing a private member of TelescopeState. Change it to
use public members provided by the new version.

It also explicitly constructs a TabloidRedis for `temp_telstate`. That's
a bit more complex than necessary at the moment (`TelescopeState()` will
use a tabloid redis backend), but future-proofs it against the default
`TelescopeState` constructor switching to the in-memory backend.